+++ src/js/language/be_NL.js
@@ -176,6 +176,7 @@
                 SI: 'Slovenië',
                 SK: 'Slowakije',
                 SM: 'San Marino',
+                TH: 'Thailand',
                 ZA: 'Zuid-Afrika'
             }
         },
@@ -238,7 +239,9 @@
                 MA: 'Marokko',
                 PK: 'Pakistan',
                 RO: 'Roemenië',
-                US: 'VS'
+                TH: 'Thailand',
+                US: 'VS',
+                VE: 'Venezuela'
             }
         },
         regexp: {