+++ src/js/language/de_DE.js
@@ -176,6 +176,7 @@
                 SI: 'Slowenien',
                 SK: 'Slowakei',
                 SM: 'San Marino',
+                TH: 'Thailand',
                 ZA: 'Südafrika'
             }
         },
@@ -238,7 +239,9 @@
                 MA: 'Marokko',
                 PK: 'Pakistan',
                 RO: 'Rumänien',
-                US: 'Vereinigte Staaten von Amerika'
+                TH: 'Thailand',
+                US: 'Vereinigte Staaten von Amerika',
+                VE: 'Venezuela'
             }
         },
         regexp: {