+++ src/js/language/es_CL.js
@@ -176,6 +176,7 @@
                 SI: 'Eslovenia',
                 SK: 'Eslovaquia',
                 SM: 'San Marino',
+                TH: 'Tailandia',
                 ZA: 'Sudáfrica'
             }
         },
@@ -238,7 +239,9 @@
                 MA: 'Marruecos',
                 PK: 'Pakistán',
                 RO: 'Rumania',
-                US: 'Estados Unidos'
+                TH: 'Tailandia',
+                US: 'Estados Unidos',
+                VE: 'Venezuela'
             }
         },
         regexp: {