+++ src/js/language/fr_FR.js
@@ -176,6 +176,7 @@
                 SI: 'Slovénie',
                 SK: 'Slovaquie',
                 SM: 'Saint-Marin',
+                TH: 'Thaïlande',
                 ZA: 'Afrique du Sud'
             }
         },
@@ -238,7 +239,9 @@
                 MA: 'Maroc',
                 PK: 'Pakistan',
                 RO: 'Roumanie',
-                US: 'USA'
+                TH: 'Thaïlande',
+                US: 'USA',
+                VE: 'Venezuela'
             }
         },
         regexp: {