+++ src/js/language/pt_BR.js
@@ -176,6 +176,7 @@
                 SI: 'Eslovênia',
                 SK: 'Eslováquia',
                 SM: 'San Marino',
+                TH: 'Tailândia',
                 ZA: 'África do Sul'
             }
         },
@@ -238,7 +239,9 @@
                 MA: 'Marrocos',
                 PK: 'Paquistão',
                 RO: 'Roménia',
-                US: 'EUA'
+                TH: 'Tailândia',
+                US: 'EUA',
+                VE: 'Venezuela'
             }
         },
         regexp: {