+++ src/js/language/ro_RO.js
@@ -176,6 +176,7 @@
                 SI: 'Slovenia',
                 SK: 'Slovacia',
                 SM: 'San Marino',
+                TH: 'Thailanda',
                 ZA: 'Africa de Sud'
             }
         },
@@ -238,7 +239,9 @@
                 MA: 'Maroc',
                 PK: 'Pakistan',
                 RO: 'Romania',
-                US: 'SUA'
+                TH: 'Thailanda',
+                US: 'SUA',
+                VE: 'Venezuela'
             }
         },
         regexp: {