+++ src/js/language/vi_VN.js
@@ -176,6 +176,7 @@
                 SI: 'Slovenia',
                 SK: 'Slovakia',
                 SM: 'San Marino',
+                TH: 'Thái Lan',
                 ZA: 'Nam Phi'
             }
         },
@@ -238,7 +239,9 @@
                 MA: 'Maroc',
                 PK: 'Pakistan',
                 RO: 'Romania',
-                US: 'Mỹ'
+                TH: 'Thái Lan',
+                US: 'Mỹ',
+                VE: 'Venezuela'
             }
         },
         regexp: {